KeypairGenerator, FriendbotFundAccount, MuxedAccount components
